locked_waker: Remove spinlock
Performance is slightly better or no difference, it makes the code
cleaner.

Removal the logic is safe. Because if waker sent but not seen by the otherside,
it's safe to cancel. If seen by the otherside and wake() and me got/sent message,
means no others competitor befor me, offset the effect.